+++ src/gui/GuiRunner.cc
@@ -60,10 +60,13 @@ class ignition::gazebo::GuiRunner::Implementation
   public: std::thread updateThread;
 };
 
+
 /////////////////////////////////////////////////
 GuiRunner::GuiRunner(const std::string &_worldName)
   : dataPtr(utils::MakeUniqueImpl<Implementation>())
 {
+  qRegisterMetaType<msgs::SerializedStepMap>();
+
   this->setProperty("worldName", QString::fromStdString(_worldName));
 
   auto win = gui::App()->findChild<ignition::gui::MainWindow *>();
@@ -91,20 +94,9 @@ GuiRunner::GuiRunner(const std::string &_worldName)
   this->RequestState();
 
   // Periodically update the plugins
-  // \todo(anyone) Move the global variables to GuiRunner::Implementation on v5
-  this->dataPtr->running = true;
-  this->dataPtr->updateThread = std::thread([&]()
-  {
-    while (this->dataPtr->running)
-    {
-      {
-        std::lock_guard<std::mutex> lock(this->dataPtr->updateMutex);
-        this->dataPtr->UpdatePlugins();
-      }
-      // This is roughly a 30Hz update rate.
-      std::this_thread::sleep_for(std::chrono::milliseconds(33));
-    }
-  });
+  QPointer<QTimer> timer = new QTimer(this);
+  connect(timer, &QTimer::timeout, this, &GuiRunner::UpdatePlugins);
+  timer->start(33);
 }
 
 /////////////////////////////////////////////////
@@ -181,9 +173,20 @@ void GuiRunner::OnState(const msgs::SerializedStepMap &_msg)
 {
   IGN_PROFILE_THREAD_NAME("GuiRunner::OnState");
   IGN_PROFILE("GuiRunner::Update");
+  // Since this function may be called from a transport thread, we push the
+  // OnStateQt function to the queue so that its called from the Qt thread. This
+  // ensures that only one thread has access to the ecm and updateInfo
+  // variables.
+  QMetaObject::invokeMethod(this, "OnStateQt", Qt::QueuedConnection,
+                            Q_ARG(msgs::SerializedStepMap, _msg));
+}
 
-  std::lock_guard<std::mutex> lock(this->dataPtr->updateMutex);
-  this->dataPtr->ecm.SetState(_msg.state());
+/////////////////////////////////////////////////
+void GuiRunner::OnStateQt(const msgs::SerializedStepMap &_msg)
+{
+  IGN_PROFILE_THREAD_NAME("Qt thread");
+  IGN_PROFILE("GuiRunner::Update");
+  this->ecm.SetState(_msg.state());
 
   // Update all plugins
   this->dataPtr->updateInfo = convert<UpdateInfo>(_msg.stats());