+++ src/gui/plugins/component_inspector/ComponentInspector.cc
@@ -17,6 +17,7 @@
 
 #include <iostream>
 #include <regex>
+#include <QColorDialog>
 #include <ignition/common/Console.hh>
 #include <ignition/common/Profiler.hh>
 #include <ignition/gui/Application.hh>
@@ -27,12 +28,15 @@
 #include "ignition/gazebo/components/Actor.hh"
 #include "ignition/gazebo/components/AngularAcceleration.hh"
 #include "ignition/gazebo/components/AngularVelocity.hh"
+#include "ignition/gazebo/components/BatterySoC.hh"
 #include "ignition/gazebo/components/CastShadows.hh"
+#include "ignition/gazebo/components/CenterOfVolume.hh"
 #include "ignition/gazebo/components/ChildLinkName.hh"
 #include "ignition/gazebo/components/Collision.hh"
 #include "ignition/gazebo/components/Factory.hh"
 #include "ignition/gazebo/components/Gravity.hh"
 #include "ignition/gazebo/components/Joint.hh"
+#include "ignition/gazebo/components/LaserRetro.hh"
 #include "ignition/gazebo/components/Level.hh"
 #include "ignition/gazebo/components/Light.hh"
 #include "ignition/gazebo/components/LightCmd.hh"
@@ -42,6 +46,7 @@
 #include "ignition/gazebo/components/LinearVelocitySeed.hh"
 #include "ignition/gazebo/components/Link.hh"
 #include "ignition/gazebo/components/MagneticField.hh"
+#include "ignition/gazebo/components/Material.hh"
 #include "ignition/gazebo/components/Model.hh"
 #include "ignition/gazebo/components/Name.hh"
 #include "ignition/gazebo/components/ParentEntity.hh"
@@ -58,7 +63,10 @@
 #include "ignition/gazebo/components/Sensor.hh"
 #include "ignition/gazebo/components/SourceFilePath.hh"
 #include "ignition/gazebo/components/Static.hh"
+#include "ignition/gazebo/components/ThreadPitch.hh"
+#include "ignition/gazebo/components/Transparency.hh"
 #include "ignition/gazebo/components/Visual.hh"
+#include "ignition/gazebo/components/Volume.hh"
 #include "ignition/gazebo/components/WindMode.hh"
 #include "ignition/gazebo/components/World.hh"
 #include "ignition/gazebo/EntityComponentManager.hh"
@@ -242,6 +250,13 @@ void ignition::gazebo::setData(QStandardItem *_item, const int &_data)
 
 //////////////////////////////////////////////////
 template<>
+void ignition::gazebo::setData(QStandardItem *_item, const Entity &_data)
+{
+  setData(_item, static_cast<int>(_data));
+}
+
+//////////////////////////////////////////////////
+template<>
 void ignition::gazebo::setData(QStandardItem *_item, const double &_data)
 {
   if (nullptr == _item)
@@ -265,6 +280,39 @@ void ignition::gazebo::setData(QStandardItem *_item, const sdf::Physics &_data)
     QVariant(_data.MaxStepSize()),
     QVariant(_data.RealTimeFactor())
   }), ComponentsModel::RoleNames().key("data"));
+}
+
+//////////////////////////////////////////////////
+template<>
+void ignition::gazebo::setData(QStandardItem *_item,
+    const sdf::Material &_data)
+{
+  if (nullptr == _item)
+    return;
+
+  _item->setData(QString("Material"),
+      ComponentsModel::RoleNames().key("dataType"));
+  _item->setData(QList({
+    QVariant(_data.Ambient().R() * 255),
+    QVariant(_data.Ambient().G() * 255),
+    QVariant(_data.Ambient().B() * 255),
+    QVariant(_data.Ambient().A() * 255),
+    QVariant(_data.Diffuse().R() * 255),
+    QVariant(_data.Diffuse().G() * 255),
+    QVariant(_data.Diffuse().B() * 255),
+    QVariant(_data.Diffuse().A() * 255),
+    QVariant(_data.Specular().R() * 255),
+    QVariant(_data.Specular().G() * 255),
+    QVariant(_data.Specular().B() * 255),
+    QVariant(_data.Specular().A() * 255),
+    QVariant(_data.Emissive().R() * 255),
+    QVariant(_data.Emissive().G() * 255),
+    QVariant(_data.Emissive().B() * 255),
+    QVariant(_data.Emissive().A() * 255)
+  }), ComponentsModel::RoleNames().key("data"));
+
+  // TODO(anyone) Only shows colors of material,
+  // need to add others (e.g., pbr)
 }
 
 //////////////////////////////////////////////////
@@ -366,6 +414,7 @@ ComponentInspector::ComponentInspector()
   : GuiSystem(), dataPtr(std::make_unique<ComponentInspectorPrivate>())
 {
   qRegisterMetaType<ignition::gazebo::ComponentTypeId>();
+  qRegisterMetaType<Entity>("Entity");
 }
 
 /////////////////////////////////////////////////
@@ -482,12 +531,7 @@ void ComponentInspector::Update(const UpdateInfo &,
     // Add component to list
     else
     {
-      // TODO(louise) Blocking here is not the best idea
-      QMetaObject::invokeMethod(&this->dataPtr->componentsModel,
-          "AddComponentType",
-          Qt::BlockingQueuedConnection,
-          Q_RETURN_ARG(QStandardItem *, item),
-          Q_ARG(ignition::gazebo::ComponentTypeId, typeId));
+      item = this->dataPtr->componentsModel.AddComponentType(typeId);
     }
 
     item->setData(QString::number(this->dataPtr->entity),
@@ -528,12 +572,29 @@ void ComponentInspector::Update(const UpdateInfo &,
       if (comp)
         setData(item, comp->Data());
     }
+    else if (typeId == components::BatterySoC::typeId)
+    {
+      auto comp = _ecm.Component<components::BatterySoC>(
+          this->dataPtr->entity);
+      if (comp)
+        setData(item, comp->Data());
+    }
     else if (typeId == components::CastShadows::typeId)
     {
       auto comp = _ecm.Component<components::CastShadows>(
           this->dataPtr->entity);
       if (comp)
         setData(item, comp->Data());
+    }
+    else if (typeId == components::CenterOfVolume::typeId)
+    {
+      auto comp = _ecm.Component<components::CenterOfVolume>(
+          this->dataPtr->entity);
+      if (comp)
+      {
+        setData(item, comp->Data());
+        setUnit(item, "m");
+      }
     }
     else if (typeId == components::ChildLinkName::typeId)
     {
@@ -550,6 +611,12 @@ void ComponentInspector::Update(const UpdateInfo &,
         setData(item, comp->Data());
         setUnit(item, "m/s\u00B2");
       }
+    }
+    else if (typeId == components::LaserRetro::typeId)
+    {
+      auto comp = _ecm.Component<components::LaserRetro>(this->dataPtr->entity);
+      if (comp)
+        setData(item, comp->Data());
     }
     else if (typeId == components::LinearAcceleration::typeId)
     {
@@ -695,6 +762,33 @@ void ComponentInspector::Update(const UpdateInfo &,
       if (comp)
         setData(item, comp->Data());
     }
+    else if (typeId == components::ThreadPitch::typeId)
+    {
+      auto comp = _ecm.Component<components::ThreadPitch>(
+          this->dataPtr->entity);
+      if (comp)
+      {
+        setData(item, comp->Data());
+        setUnit(item, "m");
+      }
+    }
+    else if (typeId == components::Transparency::typeId)
+    {
+      auto comp = _ecm.Component<components::Transparency>(
+          this->dataPtr->entity);
+      if (comp)
+        setData(item, comp->Data());
+    }
+    else if (typeId == components::Volume::typeId)
+    {
+      auto comp = _ecm.Component<components::Volume>(
+          this->dataPtr->entity);
+      if (comp)
+      {
+        setData(item, comp->Data());
+        setUnit(item, "m\u00B3");
+      }
+    }
     else if (typeId == components::WindMode::typeId)
     {
       auto comp = _ecm.Component<components::WindMode>(this->dataPtr->entity);
@@ -709,6 +803,16 @@ void ComponentInspector::Update(const UpdateInfo &,
       {
         setData(item, comp->Data());
         setUnit(item, "rad/s\u00B2");
+      }
+    }
+    else if (typeId == components::WorldAngularVelocity::typeId)
+    {
+      auto comp = _ecm.Component<components::WorldAngularVelocity>(
+          this->dataPtr->entity);
+      if (comp)
+      {
+        setData(item, comp->Data());
+        setUnit(item, "rad/s");
       }
     }
     else if (typeId == components::WorldLinearVelocity::typeId)
@@ -743,6 +847,15 @@ void ComponentInspector::Update(const UpdateInfo &,
           this->dataPtr->entity);
       if (comp)
         setData(item, comp->Data());
+    }
+    else if (typeId == components::Material::typeId)
+    {
+      auto comp = _ecm.Component<components::Material>(this->dataPtr->entity);
+      if (comp)
+      {
+        this->SetType("material");
+        setData(item, comp->Data());
+      }
     }
   }
 
@@ -790,13 +903,13 @@ bool ComponentInspector::eventFilter(QObject *_obj, QEvent *_event)
 }
 
 /////////////////////////////////////////////////
-int ComponentInspector::Entity() const
+Entity ComponentInspector::GetEntity() const
 {
   return this->dataPtr->entity;
 }
 
 /////////////////////////////////////////////////
-void ComponentInspector::SetEntity(const int &_entity)
+void ComponentInspector::SetEntity(const Entity &_entity)
 {
   // If nothing is selected, display world properties
   if (_entity == kNullEntity)
@@ -952,6 +1065,96 @@ void ComponentInspector::OnPhysics(double _stepSize, double _realTimeFactor)
     return;
   }
   this->dataPtr->node.Request(physicsCmdService, req, cb);
+}
+
+/////////////////////////////////////////////////
+void ComponentInspector::OnMaterialColor(
+  double _rAmbient, double _gAmbient, double _bAmbient, double _aAmbient,
+  double _rDiffuse, double _gDiffuse, double _bDiffuse, double _aDiffuse,
+  double _rSpecular, double _gSpecular, double _bSpecular, double _aSpecular,
+  double _rEmissive, double _gEmissive, double _bEmissive, double _aEmissive,
+  QString _type, QColor _currColor)
+{
+  // when type is not empty, open qt color dialog
+  std::string type = _type.toStdString();
+  if (!type.empty())
+  {
+    QColor newColor = QColorDialog::getColor(
+        _currColor, nullptr, "Pick a color",
+        {QColorDialog::DontUseNativeDialog, QColorDialog::ShowAlphaChannel});
+
+    // returns if the user hits cancel
+    if (!newColor.isValid())
+      return;
+
+    if (type == "ambient")
+    {
+      _rAmbient = newColor.red();
+      _gAmbient = newColor.green();
+      _bAmbient = newColor.blue();
+      _aAmbient = newColor.alpha();
+    }
+    else if (type == "diffuse")
+    {
+      _rDiffuse = newColor.red();
+      _gDiffuse = newColor.green();
+      _bDiffuse = newColor.blue();
+      _aDiffuse = newColor.alpha();
+    }
+    else if (type == "specular")
+    {
+      _rSpecular = newColor.red();
+      _gSpecular = newColor.green();
+      _bSpecular = newColor.blue();
+      _aSpecular = newColor.alpha();
+    }
+    else if (type == "emissive")
+    {
+      _rEmissive = newColor.red();
+      _gEmissive = newColor.green();
+      _bEmissive = newColor.blue();
+      _aEmissive = newColor.alpha();
+    }
+    else
+    {
+      ignerr << "Invalid material type: " << type << std::endl;
+      return;
+    }
+  }
+
+  std::function<void(const ignition::msgs::Boolean &, const bool)> cb =
+      [](const ignition::msgs::Boolean &/*_rep*/, const bool _result)
+  {
+    if (!_result)
+      ignerr << "Error setting material color configuration"
+             << " on visual" << std::endl;
+  };
+
+  msgs::Visual req;
+  req.set_id(this->dataPtr->entity);
+
+  msgs::Set(req.mutable_material()->mutable_ambient(),
+    math::Color(_rAmbient / 255.0, _gAmbient / 255.0,
+      _bAmbient / 255.0, _aAmbient / 255.0));
+  msgs::Set(req.mutable_material()->mutable_diffuse(),
+    math::Color(_rDiffuse / 255.0, _gDiffuse / 255.0,
+      _bDiffuse / 255.0, _aDiffuse / 255.0));
+  msgs::Set(req.mutable_material()->mutable_specular(),
+    math::Color(_rSpecular / 255.0, _gSpecular / 255.0,
+      _bSpecular / 255.0, _aSpecular / 255.0));
+  msgs::Set(req.mutable_material()->mutable_emissive(),
+    math::Color(_rEmissive / 255.0, _gEmissive / 255.0,
+      _bEmissive / 255.0, _aEmissive / 255.0));
+
+  auto materialCmdService = "/world/" + this->dataPtr->worldName
+      + "/visual_config";
+  materialCmdService = transport::TopicUtils::AsValidTopic(materialCmdService);
+  if (materialCmdService.empty())
+  {
+    ignerr << "Invalid material command service topic provided" << std::endl;
+    return;
+  }
+  this->dataPtr->node.Request(materialCmdService, req, cb);
 }
 
 /////////////////////////////////////////////////