+++ src/gui/plugins/component_inspector/ComponentInspector.hh
@@ -115,6 +115,13 @@ namespace gazebo
   template<>
   void setData(QStandardItem *_item, const std::ostream &_data);
 
+  /// \brief Specialized to set material data.
+  /// \param[in] _item Item whose data will be set.
+  /// \param[in] _data Data to set.
+  template<>
+  void setData(QStandardItem *_item, const sdf::Material &_data);
+
+
   /// \brief Set the unit of a given item.
   /// \param[in] _item Item whose unit will be set.
   /// \param[in] _unit Unit to be displayed, such as 'm' for meters.
@@ -164,8 +171,8 @@ namespace gazebo
 
     /// \brief Entity
     Q_PROPERTY(
-      int entity
-      READ Entity
+      Entity entity
+      READ GetEntity
       WRITE SetEntity
       NOTIFY EntityChanged
     )
@@ -260,6 +267,35 @@ namespace gazebo
     public: Q_INVOKABLE void OnPhysics(double _stepSize,
         double _realTimeFactor);
 
+    // \brief Callback in Qt thread when material color changes for a visual
+    /// \param[in] _rAmbient ambient red
+    /// \param[in] _gAmbient ambient green
+    /// \param[in] _bAmbient ambient blue
+    /// \param[in] _aAmbient ambient alpha
+    /// \param[in] _rDiffuse diffuse red
+    /// \param[in] _gDiffuse diffuse green
+    /// \param[in] _bDiffuse diffuse blue
+    /// \param[in] _aDiffuse diffuse alpha
+    /// \param[in] _rSpecular specular red
+    /// \param[in] _gSpecular specular green
+    /// \param[in] _bSpecular specular blue
+    /// \param[in] _aSpecular specular alpha
+    /// \param[in] _rEmissive emissive red
+    /// \param[in] _gEmissive emissive green
+    /// \param[in] _bEmissive emissive blue
+    /// \param[in] _aEmissive emissive alpha
+    /// \param[in] _type if type is not empty, opens QColorDialog.
+    /// The possible types are ambient, diffuse, specular, or emissive.
+    /// \param[in] _currColor used for QColorDialog to show the current color
+    /// in the open dialog.
+    public: Q_INVOKABLE void OnMaterialColor(
+      double _rAmbient, double _gAmbient, double _bAmbient,
+      double _aAmbient, double _rDiffuse, double _gDiffuse,
+      double _bDiffuse, double _aDiffuse, double _rSpecular,
+      double _gSpecular, double _bSpecular, double _aSpecular,
+      double _rEmissive, double _gEmissive, double _bEmissive,
+      double _aEmissive, QString _type, QColor _currColor);
+
     /// \brief Get whether the entity is a nested model or not
     /// \return True if the entity is a nested model, false otherwise
     public: Q_INVOKABLE bool NestedModel() const;
@@ -272,11 +308,11 @@ namespace gazebo
 
     /// \brief Get the entity currently inspected.
     /// \return Entity ID.
-    public: Q_INVOKABLE int Entity() const;
+    public: Q_INVOKABLE Entity GetEntity() const;
 
     /// \brief Set the entity currently inspected.
     /// \param[in] _entity Entity ID.
-    public: Q_INVOKABLE void SetEntity(const int &_entity);
+    public: Q_INVOKABLE void SetEntity(const Entity &_entity);
 
     /// \brief Notify that entity has changed.
     signals: void EntityChanged();