+++ src/gui/plugins/shapes/Shapes.cc
@@ -78,6 +78,7 @@ void Shapes::OnMode(const QString &_mode)
   std::transform(modelSdfString.begin(), modelSdfString.end(),
                  modelSdfString.begin(), ::tolower);
 
+  // TODO(anyone) when porting to v5 add <material> tag to capsule & ellipsoid
   if (modelSdfString == "box")
   {
     modelSdfString = std::string("<?xml version=\"1.0\"?>"
@@ -109,6 +110,11 @@ void Shapes::OnMode(const QString &_mode)
                                              "<size>1 1 1</size>"
                                            "</box>"
                                          "</geometry>"
+                                         "<material>"
+                                           "<ambient>0.3 0.3 0.3 1</ambient>"
+                                           "<diffuse>0.7 0.7 0.7 1</diffuse>"
+                                           "<specular>1 1 1 1</specular>"
+                                         "</material>"
                                        "</visual>"
                                      "</link>"
                                    "</model>"
@@ -145,6 +151,11 @@ void Shapes::OnMode(const QString &_mode)
                                              "<radius>0.5</radius>"
                                            "</sphere>"
                                          "</geometry>"
+                                         "<material>"
+                                           "<ambient>0.3 0.3 0.3 1</ambient>"
+                                           "<diffuse>0.7 0.7 0.7 1</diffuse>"
+                                           "<specular>1 1 1 1</specular>"
+                                         "</material>"
                                        "</visual>"
                                      "</link>"
                                    "</model>"
@@ -183,6 +194,11 @@ void Shapes::OnMode(const QString &_mode)
                                              "<length>1.0</length>"
                                            "</cylinder>"
                                          "</geometry>"
+                                         "<material>"
+                                           "<ambient>0.3 0.3 0.3 1</ambient>"
+                                           "<diffuse>0.7 0.7 0.7 1</diffuse>"
+                                           "<specular>1 1 1 1</specular>"
+                                         "</material>"
                                        "</visual>"
                                      "</link>"
                                    "</model>"